refactor(api): renames contactRequests file to "contactRequests"
