+++ api/src/routes/contacts/handlers/listReceivedContactRequests.ts
@@ -1,6 +1,6 @@
 import { UnauthorizedError } from "../../../errors/client";
-import { ContactRequest } from "../../../services/drizzle-database/schemas/contact";
-import { listReceivedContactRequests } from "../../../services/drizzle-database/queries/contact/listReceivedContactRequests";
+import { ContactRequest } from "../../../services/drizzle-database/schemas/contactRequest";
+import { listReceivedContactRequests } from "../../../services/drizzle-database/queries/contactRequest/listReceivedContactRequests";
 import { User } from "../../../services/drizzle-database/schemas/auth";
 import { Database } from "../../../services/drizzle-database/client";
