+++ api/src/services/drizzle-database/queries/contactRequest/createContactRequest.ts
@@ -1,6 +1,6 @@
 import { randomUUID } from "crypto";
 import { Database } from "../../client";
-import { ContactRequest, contactRequest } from "../../schemas/contact";
+import { ContactRequest, contactRequest } from "../../schemas/contactRequest";
 import { logger } from "../../../../logger";
 import { InternalServerError } from "../../../../errors/server";
