+++ api/src/services/drizzle-database/queries/contactRequest/listSentContactRequests.ts
@@ -1,31 +1,30 @@
 import { Database } from "../../client";
-import { ContactRequest, contactRequest } from "../../schemas/contact";
+import { ContactRequest, contactRequest } from "../../schemas/contactRequest";
 import { logger } from "../../../../logger";
 import { InternalServerError } from "../../../../errors/server";
 import { and, eq, isNull } from "drizzle-orm";
 import { User, user } from "../../schemas/auth";
 
-export const listReceivedContactRequests = async (
+export const listSentContactRequests = async (
   database: Database,
-  { receiverUserId }: Pick<ContactRequest, "receiverUserId">,
+  { senderUserId }: Pick<ContactRequest, "senderUserId">,
 ): Promise<{ contactRequest: ContactRequest; user: User }[]> => {
   const rows = await database
     .select()
     .from(contactRequest)
-    .innerJoin(user, eq(contactRequest.senderUserId, user.id))
+    .innerJoin(user, eq(contactRequest.receiverUserId, user.id))
     .where(
       and(
-        eq(contactRequest.receiverUserId, receiverUserId),
+        eq(contactRequest.senderUserId, senderUserId),
         eq(contactRequest.status, "PENDING"),
         isNull(contactRequest.deletedAt),
       ),
     )
     .catch((error) => {
-      logger.error(error, "Error while listing received contact requests.");
+      logger.error(error, "Error while listing sent contact requests.");
       throw new InternalServerError({
-        code: "unknown-error-listing-received-contact-requests",
-        message:
-          "Unknown error when trying to listing received contact requests",
+        code: "unknown-error-listing-sent-contact-requests",
+        message: "Unknown error when trying to listing sent contact requests",
       });
     });
